Don't install net-tools on opensuse

+++ scripts/bk_tests/bk_container_prep.sh
@@ -5,7 +5,7 @@ if [ -f /etc/debian_version ]; then
   apt-get update -y && apt-get install -y net-tools iproute2
   mkdir -p /etc/network/interfaces.d
   touch /etc/network/interfaces
-else
+elif [ -f /etc/redhat-release ]; then
   yum install -y net-tools
 fi
